Trivia and crossword game
First it askes a trivia question, then if you get it wrong it will give you a different random trivia question. if you get it correct then it will give you a crossword and you have to type in the index values of all the letters in the word to get it correct. then it gives you another trivia, and when you get 10 points or finish all the trivia you win.